app/requirements.txt
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
python-dotenv==1.1.1